+++ noodles-csi/CHANGELOG.md
@@ -28,6 +28,8 @@
     It isn't clear what `aux` is supposed to be used for, so noodles-csi
     assumes if it's set, it's a tabix header.
 
+  * csi/index: Optimize chunks from `Index::query`.
+
   * csi/index/reference_sequence: Increase the visibility of `Builder`.
 
   * csi/index/reference_sequence/bin: Increase the visibility of `Builder`.